Sandra's update
 Saturday, April 3, 2021 4:53 AM

Here is my part for the assignment. I think after the weekend I am going to have to make it neater but I thought it would be best to share with you guys to get an input, like if I misunderstood something. By the way, for some reason I cannot write into textfiles outside the buld-assignment-desktop folder so if you want to test if it works for you go a head and change the path in the code. I don't know why it is denying me permission but I think it may do with something with Makefile in there. I am talking off my butt but if you guys somehow do it well yay, it's probably my OS denying me permission or something.